Split rocsp-tool into files, add some tests
