+++ cmd/rocsp-tool/client.go
@@ -1,97 +1,24 @@
 package notmain
 
 import (
-	"bytes"
 	"context"
-	"crypto/x509/pkix"
 	"database/sql"
-	"encoding/asn1"
 	"fmt"
 	"io/ioutil"
 	"log"
 	"math/rand"
 	"strings"
-	"sync"
 	"sync/atomic"
 	"time"
 
 	"github.com/jmhodges/clock"
 	capb "github.com/letsencrypt/boulder/ca/proto"
 	"github.com/letsencrypt/boulder/core"
-	"github.com/letsencrypt/boulder/issuance"
 	"github.com/letsencrypt/boulder/rocsp"
 	"github.com/letsencrypt/boulder/sa"
 	"github.com/letsencrypt/boulder/test/ocsp/helper"
 	"golang.org/x/crypto/ocsp"
 )
-
-type shortIDIssuer struct {
-	*issuance.Certificate
-	subject      pkix.RDNSequence
-	shortID      byte
-	issuerID     issuance.IssuerID
-	issuerNameID issuance.IssuerNameID
-}
-
-func loadIssuers(input map[string]int) ([]shortIDIssuer, error) {
-	var issuers []shortIDIssuer
-	for issuerFile, shortID := range input {
-		if shortID > 255 || shortID < 0 {
-			return nil, fmt.Errorf("invalid shortID %d (must be byte)", shortID)
-		}
-		cert, err := issuance.LoadCertificate(issuerFile)
-		if err != nil {
-			return nil, fmt.Errorf("reading issuer: %w", err)
-		}
-		var subject pkix.RDNSequence
-		_, err = asn1.Unmarshal(cert.Certificate.RawSubject, &subject)
-		if err != nil {
-			return nil, fmt.Errorf("parsing issuer.RawSubject: %w", err)
-		}
-		var shortID byte = byte(shortID)
-		for _, issuer := range issuers {
-			if issuer.shortID == shortID {
-				return nil, fmt.Errorf("duplicate shortID in config file: %d (for %q and %q)", shortID, issuer.subject, subject)
-			}
-			if !issuer.IsCA {
-				return nil, fmt.Errorf("certificate for %q is not a CA certificate", subject)
-			}
-		}
-		issuers = append(issuers, shortIDIssuer{
-			Certificate:  cert,
-			subject:      subject,
-			shortID:      shortID,
-			issuerID:     cert.ID(),
-			issuerNameID: cert.NameID(),
-		})
-	}
-	return issuers, nil
-}
-
-func findIssuerByID(longID int64, issuers []shortIDIssuer) (*shortIDIssuer, error) {
-	for _, iss := range issuers {
-		if iss.issuerNameID == issuance.IssuerNameID(longID) || iss.issuerID == issuance.IssuerID(longID) {
-			return &iss, nil
-		}
-	}
-	return nil, fmt.Errorf("no issuer found for an ID in certificateStatus: %d", longID)
-}
-
-func findIssuerByName(resp *ocsp.Response, issuers []shortIDIssuer) (*shortIDIssuer, error) {
-	var responder pkix.RDNSequence
-	_, err := asn1.Unmarshal(resp.RawResponderName, &responder)
-	if err != nil {
-		return nil, fmt.Errorf("parsing resp.RawResponderName: %w", err)
-	}
-	var responders strings.Builder
-	for _, issuer := range issuers {
-		fmt.Fprintf(&responders, "%s\n", issuer.subject)
-		if bytes.Equal(issuer.RawSubject, resp.RawResponderName) {
-			return &issuer, nil
-		}
-	}
-	return nil, fmt.Errorf("no issuer found matching OCSP response for %s. Available issuers:\n%s\n", responder, responders.String())
-}
 
 type client struct {
 	issuers       []shortIDIssuer
@@ -99,56 +26,6 @@ type client struct {
 	db            *sql.DB // optional
 	ocspGenerator capb.OCSPGeneratorClient
 	clk           clock.Clock
-}
-
-type inflight struct {
-	sync.RWMutex
-	items map[uint64]struct{}
-}
-
-func newInflight() *inflight {
-	return &inflight{
-		items: make(map[uint64]struct{}),
-	}
-}
-
-func (i *inflight) add(n uint64) {
-	i.Lock()
-	defer i.Unlock()
-	i.items[n] = struct{}{}
-}
-
-func (i *inflight) remove(n uint64) {
-	i.Lock()
-	defer i.Unlock()
-	delete(i.items, n)
-}
-
-func (i *inflight) len() int {
-	i.RLock()
-	defer i.RUnlock()
-	return len(i.items)
-}
-
-// min returns the numerically smallest key inflight. If nothing is inflight,
-// it returns 0. Note: this takes O(n) time in the number of keys and should
-// be called rarely.
-func (i *inflight) min() uint64 {
-	i.RLock()
-	defer i.RUnlock()
-	if len(i.items) == 0 {
-		return 0
-	}
-	var min uint64
-	for k := range i.items {
-		if min == 0 {
-			min = k
-		}
-		if k < min {
-			min = k
-		}
-	}
-	return min
 }
 
 // processResult represents the result of attempting to sign and store status
