+++ cmd/rocsp-tool/client_test.go
@@ -1,80 +1,66 @@
 package notmain
 
 import (
-	"encoding/hex"
-	"strings"
+	"context"
+	"io/ioutil"
 	"testing"
+	"time"
 
+	"github.com/go-redis/redis/v8"
+	"github.com/jmhodges/clock"
+	"github.com/letsencrypt/boulder/cmd"
+	"github.com/letsencrypt/boulder/rocsp"
 	"github.com/letsencrypt/boulder/test"
-	"golang.org/x/crypto/ocsp"
 )
 
-func TestLoadIssuers(t *testing.T) {
-	input := map[string]int{
-		"../../test/hierarchy/int-e1.cert.pem": 23,
-		"../../test/hierarchy/int-r3.cert.pem": 99,
+func makeClient() (*rocsp.WritingClient, clock.Clock) {
+	CACertFile := "../test/redis-tls/minica.pem"
+	CertFile := "../test/redis-tls/boulder/cert.pem"
+	KeyFile := "../test/redis-tls/boulder/key.pem"
+	tlsConfig := cmd.TLSConfig{
+		CACertFile: &CACertFile,
+		CertFile:   &CertFile,
+		KeyFile:    &KeyFile,
 	}
-	output, err := loadIssuers(input)
+	tlsConfig2, err := tlsConfig.Load()
 	if err != nil {
-		t.Fatal(err)
+		panic(err)
 	}
 
-	var e1 *shortIDIssuer
-	var r3 *shortIDIssuer
-
-	for i, v := range output {
-		if strings.Contains(v.Certificate.Subject.String(), "E1") {
-			e1 = &output[i]
-		}
-		if strings.Contains(v.Certificate.Subject.String(), "R3") {
-			r3 = &output[i]
-		}
-	}
-
-	test.AssertEquals(t, e1.Subject.String(), "CN=(TEST) Elegant Elephant E1,O=Boulder Test,C=XX")
-	test.AssertEquals(t, r3.Subject.String(), "CN=(TEST) Radical Rhino R3,O=Boulder Test,C=XX")
-	test.AssertEquals(t, e1.shortID, uint8(23))
-	test.AssertEquals(t, r3.shortID, uint8(99))
+	rdb := redis.NewClusterClient(&redis.ClusterOptions{
+		Addrs:     []string{"10.33.33.2:4218"},
+		Username:  "unittest-rw",
+		Password:  "824968fa490f4ecec1e52d5e34916bdb60d45f8d",
+		TLSConfig: tlsConfig2,
+	})
+	clk := clock.NewFake()
+	return rocsp.NewWritingClient(rdb, 500*time.Millisecond, clk), clk
 }
 
-func TestFindIssuer(t *testing.T) {
-	input := map[string]int{
-		"../../test/hierarchy/int-e1.cert.pem": 23,
+func TestStoreResponse(t *testing.T) {
+	redisClient, clk := makeClient()
+
+	response, err := ioutil.ReadFile("testdata/ocsp.response")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	issuers, err := loadIssuers(map[string]int{
 		"../../test/hierarchy/int-r3.cert.pem": 99,
-	}
-	issuers, err := loadIssuers(input)
+	})
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	elephant, err := hex.DecodeString("3049310b300906035504061302585831153013060355040a130c426f756c6465722054657374312330210603550403131a28544553542920456c6567616e7420456c657068616e74204531")
-	if err != nil {
-		t.Fatal(err)
-	}
-	rhino, err := hex.DecodeString("3046310b300906035504061302585831153013060355040a130c426f756c64657220546573743120301e06035504031317285445535429205261646963616c205268696e6f205233")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	ocspResp := &ocsp.Response{
-		RawResponderName: elephant,
+	cl := client{
+		issuers:       issuers,
+		redis:         redisClient,
+		db:            nil,
+		ocspGenerator: nil,
+		clk:           clk,
 	}
 
-	issuer, err := findIssuerByName(ocspResp, issuers)
-	if err != nil {
-		t.Fatalf("couldn't find issuer: %s", err)
-	}
-
-	test.AssertEquals(t, issuer.shortID, uint8(23))
-
-	ocspResp = &ocsp.Response{
-		RawResponderName: rhino,
-	}
-
-	issuer, err = findIssuerByName(ocspResp, issuers)
-	if err != nil {
-		t.Fatalf("couldn't find issuer: %s", err)
-	}
-
-	test.AssertEquals(t, issuer.shortID, uint8(99))
+	ttl := time.Hour
+	err = cl.storeResponse(context.Background(), response, &ttl)
+	test.AssertNotError(t, err, "storing response")
 }
